+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.2] - 2024-05-09
+
+### Fixed
+- problem with step loading using `--replay` option
+- made access to `workflow.output` more robust by using hash with indifferent access
+
+[0.1.2]: https://github.com/Shopify/roast/releases/tag/v0.1.2
+
 ## [0.1.1] - 2024-05-09
 
 ### Added